Crate swc_error_reporters

Modules

handler

Structs

ErrorEmitter

GraphicalReportHandler

A ReportHandler that displays a given Report in a quasi-graphical way, using terminal colors, unicode drawing characters, and other such things.

GraphicalTheme

Theme used by GraphicalReportHandler to render fancy Diagnostic reports.

TWithDiagnosticArray

A wrapper around a value that also contains a list of diagnostics. It helps swc error system migrate to the new error reporting system.

Traits

ToPrettyDiagnostic

Functions

convert_span

to_pretty_source_code