Crate swc_error_reporters

Modules

handler

json_emitter

Structs

GraphicalReportHandler

A ReportHandler that displays a given Report in a quasi-graphical way, using terminal colors, unicode drawing characters, and other such things.

GraphicalTheme

Theme used by GraphicalReportHandler to render fancy Diagnostic reports.

PrettyEmitter

PrettyEmitterConfig

Functions

to_miette_source_code

to_miette_span