Crate swc_error_reporters

Modules

handler

Structs

GraphicalReportHandler

A ReportHandler that displays a given Report in a quasi-graphical way, using terminal colors, unicode drawing characters, and other such things.

GraphicalTheme

Theme used by GraphicalReportHandler to render fancy Diagnostic reports.

PrettyEmitter

PrettyEmitterConfig